Add basic push support.
The manifest I just took from push sample and modified it for the mothership app.
The MainApplication starts the take off for the airship lib.  It then calls enable push.  We
should think of a way to ask if push should be enabled.  Ill talk to Jeff about best pratices for this.

I did not include the real airship properties files because they have app secrets in them.  Probably need to make
this repo private.